feat(rivet+rivetc): support new syntax for initializing arrays (#51)
New syntax is added to create arrays and dynamic arrays. This new syntax allows you to define a value that will be used by default, as well as define the length and capacity of the array (if it is dynamic, of course).

The previous syntax was to use the builtin function `dyn_array`; with this change, this function is no longer necessary and is removed from the compiler entirely.